requirements.txt
google-api-python-client==2.120.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
pytz==2023.3
beautifulsoup4==4.12.2